+++ packing/docs-init.js
@@ -1,0 +1,9 @@
+/*
+  From -
+     https://realworldjs.medium.com/automatic-documentation-for-javascript-projects-readme-jsdoc-mermaid-86b86be9b28d
+*/
+
+/* eslint-disable */
+window.addEventListener("load", () =>
+  mermaid.initialize({ startOnLoad: true })
+);