+++ src/pxlNav/pxlBase.js
@@ -19,6 +19,7 @@ import { Environment } from './Environment.js';
 
 // HTML DOM; HUD / GUI
 import { GUI } from './gui/GUI.js';
+import { HUD } from './gui/HUD.js';
 
 // Camera and Camera Sequencing Automation
 import { Camera } from './cam/Camera.js';
@@ -41,6 +42,6 @@ import { Video } from './media/Video.js';
 export const pxlBase = { 
   Utils, FileIO, QualityController, 
   CookieManager, Timer, User, Device, 
-  Colliders, Animation, Environment, GUI, 
+  Colliders, Animation, Environment, GUI, HUD, 
   Camera, AutoCamera, Extensions, 
   MusicUtils, Audio, Video };